Debug - Enforce UTC timezone
It seems the tests are failing on the CI because the CI runs on UTC.
My local setup runs on Europe/Paris and prod is UTC too.

The issue is that I can't consistantly reproduce tests failing with UTC
locally but they seems to consistantly fail on CI.

I don't like setting the timezone explicitely like this. This is
(hopefully) a temporary situation until I better investigate.